\name{NetData}
\alias{NetData}
\title{ Data for Social Network Analysis labs }
\description{ 
  Various Network data sets that can be used for Dan McFarland's network analysis labs. Labs
  are available at:
  
  ~\\\url{http://sna.stanford.edu/rlabs.php}
  
Larger comprehensive data sets are publicly available, contact Daniel A. McFarland for details.
  
}

\usage{NetData}
\references{Krackhardt, D. (1987). Cognitive social structures. Social Networks. 9, 109-134.}
\author{Dan McFarland \email{dmcfarla@stanford.edu}, 
Solomon Messing \email{messing~at~stanford...},
Mike Nowak \email{michael.nowak@gmail.com},
Sean J Westwood \email{seanjw@stanford.edu}
}

\examples{
data(kracknets, package = "NetData")
data('studentnets.M182', package = "NetData")
data('studentnets.S641', package = "NetData")
data('studentnets.magact96.97.98', package = "NetData")
data('studentnets.m173', package="NetData")
data('studentnets.ergm173', package = "NetData")
## Examples available at:
## http://sna.stanford.edu/rlabs.php
}
\keyword{ ~sna, NetData }

