\name{NetData}
\alias{NetData}
\title{ Data for Social Network Analysis labs }
\description{ 
  Various Network data sets that can be used for Dan McFarland's network analysis labs. Labs
  will be made available in future releases of this package.
}

\usage{}

\arguments{}
\value{}
\references{Krackhardt, D. (1987). Cognitive social structures. Social Networks. 9, 109-134.}
\author{Dan McFarland \email{dmcfarla@stanford.edu}, 
Solomon Messing \email{messing~at~stanford...},
Mike Nowak \email{michael.nowak@gmail.com},
Sean J Westwood \email{seanjw@stanford.edu}
}


\examples{
data(kracknets, package = "NetData")
data(studentnets.M182, package = "NetData")
data(studentnets.S641, package = "NetData")
data(studentnets.magact96.97.98, package = "NetData")
data(studentnets.m173, package="NetData")
data(studentnets.ergm173, package = "NetData")

}
\keyword{ ~sna, NetData }

